requirement.txt
openai
fastapi 
uvicorn
aiofiles
akshare
python-dotenv